+++ Card/src/main/java/com/paypal/android/card/CardClient.kt
@@ -75,11 +75,7 @@ class CardClient internal constructor(
             approveOrderListener?.onApproveOrderThreeDSecureWillLaunch()
 
             // launch the 3DS flow
-            val urlScheme = cardRequest.threeDSecureRequest?.let {
-                val returnUrl = it.returnUrl
-                val uri = Uri.parse(returnUrl)
-                uri.scheme
-            }
+            val urlScheme = cardRequest.threeDSecureRequest.run { Uri.parse(returnUrl).scheme }
             val approveOrderMetadata =
                 ApproveOrderMetadata(cardRequest.orderID, response.paymentSource)
             val options = BrowserSwitchOptions()