+++ Card/src/main/java/com/paypal/android/card/CardRequest.kt
@@ -5,5 +5,5 @@ import com.paypal.android.card.threedsecure.ThreeDSecureRequest
 data class CardRequest(
     val orderID: String,
     val card: Card,
-    var threeDSecureRequest: ThreeDSecureRequest? = null
+    var threeDSecureRequest: ThreeDSecureRequest
 )